Adds auth0

+++ server.js
@@ -18,6 +18,7 @@ db.once('open', function(){
 });
 
 const PORT = process.env.PORT || 3002;
+const verifyUser = require('./auth');
 
 app.get('/', (req, res) => {
   res.send('server request received');
@@ -27,24 +28,37 @@ app.get('/books', handleGetBooks);
 app.post('/books', handlePostBooks);
 app.delete('/books/:id', handleDeleteBooks);
 app.put('/books/:id', handlePutBooks);
+app.get('./user', handleGetUser);
+
+function handleGetUser(req, res) {
+  verifyUser(req, (err, user)=> {
+    if(err) {
+      console.log(err);
+      res.send('invalid token');
+    } else { res.send(user); }
+  });
+}
 
 const Book = require('./models/book');
 
 async function handleGetBooks(req, res) {
-  let queryObject = {};
-  if(req.query.email) {
-    queryObject = {email: req.query.email};
-  }
-  try {
-    let books = await Book.find(queryObject);
-    if (books.length > 0) {
-      res.status(200).send(books);
+  verifyUser(req, async (err, user)=> {
+    if(err) {
+      console.log(err);
+      res.send('invalid token');
     } else {
-      res.status(404).send('No books found');
+      try {
+        let books = await Book.find({email: user.email});
+        if (books.length > 0) {
+          res.status(200).send(books);
+        } else {
+          res.status(404).send('No books found');
+        }
+      } catch(err) {
+        res.status(500).send('Server error');
+      }
     }
-  } catch(err) {
-    res.status(500).send('Server error');
-  }
+  });
 }
 
 async function handlePostBooks(req, res) {
@@ -57,13 +71,25 @@ async function handlePostBooks(req, res) {
 }
 
 async function handleDeleteBooks(req, res) {
-  let id = req.params.id;
-  try {
-    await Book.findByIdAndDelete(id);
-    res.status(200).send('Book deleted');
-  } catch(err) {
-    res.status(404).send('Unable to delete '+id);
-  }
+  verifyUser(req, async (err, user) => {
+    if (err) {
+      console.error(err);
+      res.send('invalid token');
+    } else {
+      let id = req.params.id;
+      try {
+        const book = await Book.findOne({_id: id, email: user.emali});
+        if (!book) {
+          res.status(400).send('Unable to find book');
+        } else {
+          await Book.findByIdAndDelete(id);
+          res.status(200).send('Book deleted');
+        }
+      } catch(err) {
+        res.status(404).send('Unable to delete '+id);
+      }
+    }
+  });
 }
 
 async function handlePutBooks(req, res) {
